+++ changelog.md
@@ -62,6 +62,7 @@ This version is not yet released. If you are reading this on the website, then t
   - Add completions
   - Add diagnostics
   - Add semantic highlighting
+  - Inline function hovering
 ### Website
 - Add [Tutorial Introduction](https://uiua.org/tutorial/introduction)
 - Add [primitives.json](https://uiua.org/primitives.json) for use with tooling